refacto(library): split mjml element
